+++ lib/instructor_lite/adapters/gemini.ex
@@ -67,7 +67,7 @@ defmodule InstructorLite.Adapters.Gemini do
                          ],
                          model: [
                            type: :string,
-                           default: "gemini-2.5-flash-lite",
+                           default: "gemini-3.5-flash-lite",
                            doc:
                              "Gemini [model](https://ai.google.dev/gemini-api/docs/models/gemini)"
                          ]